requirements.txt
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.55.0
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
